Fixed ADDL-based schema generation
git-svn-id: https://activeobjects.dev.java.net/svn/activeobjects/trunk@60 69e892e8-b12f-0410-b7e1-b50061bc6545

+++ ActiveObjects/src/net/java/ao/DatabaseProvider.java
@@ -98,7 +98,7 @@ public abstract class DatabaseProvider implements IDatabaseProvider {
 				return "FLOAT";
 				
 			case Types.INTEGER:
-				return "FLOAT";
+				return "INTEGER";
 				
 			case Types.LONGVARBINARY:
 				return "LONGVARBINARY";
